refactor: split admin config apply/validate engine into dedicated services
Splits ConfigApplyService into an apply-only service and a new
ConfigValidateService, moving shared logic (dependency ordering, manifest
parsing, scratch handling, schema/uniqueness validation) into a new
ConfigManifestSupport helper, and the JSON codec helpers out of
ConfigResourceController into a neutral ConfigEntityCodec util. Fixes a
false "shared" javadoc on validateSchema (applySchema now actually calls
it instead of duplicating it) and an inverted service-to-controller
dependency.
